build: fix Mac clang-17 compiler warnings
- Add missing fmt formatter for ParquetCipher::type
- Fix implicit conversion/type deduction warning for int64_t in writer tests
- Fix -Wdangling-gsl by persisting temporary PlanNodeStats map
- Initialize uninitialized Operator pointers to nullptr in OrderByTest

+++ bolt/dwio/parquet/tests/writer/ParquetWriterTest.cpp
@@ -360,7 +360,10 @@ TEST_F(ParquetWriterTest, constantToArrow) {
            kRows,
            leafPool_.get()),
        BaseVector::createConstant(
-           BIGINT(), 10000000000L, kRows, leafPool_.get()),
+           BIGINT(),
+           static_cast<int64_t>(10000000000L),
+           kRows,
+           leafPool_.get()),
        BaseVector::createConstant(DOUBLE(), 1000.0, kRows, leafPool_.get())});
   std::string parquetPath = tempPath_->path + "/constantToArrow.parquet";
   assertWrite(parquetPath, kRows, schema, data);
